APP/backend/requirements.txt
azure-identity==1.13.0
werkzeug==2.3.7
quart==0.18.4
langchain==0.0.278
openai[datalib]==0.27.8
tiktoken==0.4.0
azure-search-documents==11.4.0b6
azure-storage-blob==12.14.1
azure-cosmos==4.5.0
uvicorn[standard]==0.23.2
aiohttp==3.8.5
azure-monitor-opentelemetry==1.0.0b15
opentelemetry-instrumentation-asgi==0.40b0
opentelemetry-instrumentation-requests==0.40b0
opentelemetry-instrumentation-aiohttp-client==0.40b0
python-dotenv==1.0.0
pypdf==3.9.0
docx2txt>=0.8
azure-ai-formrecognizer==3.2.1
tenacity==8.2.2
redis==5.0.0
redisvl==0.0.3
selenium==4.12.0
unstructured==0.10.11
python-docx==0.8.11